Rewritten all Ajax validation calls to be handed asynchronously to prevent the FormBuilder to block the UI while validating.
This also mean that we write out the renderer settings on the form instead of
retrieving it from the server. Since a form is always validated and the
Dependency feature also relies on reading renderer settings, this extra html
payload is less that having to make the seperate ajax call and makes the code
much simpler.