FormRenderer is no longer a static class (finally!!)
Refactored all static FormRenderer methods into our abstract FormRendererBase,
which is now named FormRenderer as well.

Deleted FormOptions and moved the option properties directly to FormRenderer.

Expanded ValidationSummary to be more flexible in regards to if a
ValidationSummary should be shown and where on the page it should show. Also
preventing multiple validation errors for the same field from showing, now we
only show the first message per field.