Better default support for localized Fields, Validation Message etc.
It is now possible to add attributes without defining a string which will then
be resolved at runtime based on the Form name + Field name + Attribute type
using HttpContext.GetGlobalResourceObject.

It is also possible to pass a T(...) value into an attribute which will also be
resolved at runtime using HttpContext.GetGlobalResourceObject. The use of ${ is
discouraged since it ties to C1's own Resource System which should mainly be
used to localize the C1 Console.