Changed the way Hidden Fields work
Until now, any public attribute without a FieldLabel attribute was considered a
hidden form field. This had some inconvenient sideeffects on more complex POCO
based solutions and a ExcluteField attribte was introduced in changeset
2784a25fef05.

Over time this became unmanageble as well, and based on how often Hidden Fields
are actually used it has now been changed so public properties has to be marked
with a new Hidden Field attribute to be rendered as a hidden form field.