Major refactoring of localization
Optimizations of Ajax validation, so the language the form was requested with
is persisted when making Ajax calls, so we're sure to show the correct
validation messages etc.

The use of C1's own localization for managing translations of frontend-related
strings has been moved to internal Resx files, since we mainly see them as
Console related localization.

A new ResourceFacade has been added to manage how resources are loaded into the
FormBuilder. That means that the default Resx files can be replaced with a
seperate ResourceManager for providing strings, and if the ResourceManager
provides a writer, its also possible to manage strings for various parts of the
FormBuilder through the console.

Label, Help, PlaceholderText and Validation messages have been updated to first
look for a string in via the ResourceFacade using a key based on the Form and
Field name and Validation type, before falling back to using the actual value
provided by the Form Provider. This allows for scenarios where ie. POCO forms
have values set by the Developer in source code, but which can later be
overridded by an editor using the ResourceFacade.